+++ src/handlers/test-handlers-enhanced/index.ts
@@ -5,6 +5,7 @@ import { testSubtaskOperations } from "./subtask-tests.js";
 import { testLabelOperations } from "./label-tests.js";
 import { testBulkOperations } from "./bulk-tests.js";
 import { testQuickAddOperations } from "./quick-add-tests.js";
+import { testDurationAndReopenOperations } from "./duration-reopen-tests.js";
 import { TestSuite, ComprehensiveTestReport } from "./types.js";
 
 export async function handleTestAllFeaturesEnhanced(
@@ -19,6 +20,7 @@ export async function handleTestAllFeaturesEnhanced(
   suites.push(await testSubtaskOperations(todoistClient));
   suites.push(await testLabelOperations(todoistClient));
   suites.push(await testBulkOperations(todoistClient));
+  suites.push(await testDurationAndReopenOperations(todoistClient));
 
   // Run quick add tests if API token is provided
   if (apiToken) {
@@ -61,3 +63,4 @@ export { testSubtaskOperations } from "./subtask-tests.js";
 export { testLabelOperations } from "./label-tests.js";
 export { testBulkOperations } from "./bulk-tests.js";
 export { testQuickAddOperations } from "./quick-add-tests.js";
+export { testDurationAndReopenOperations } from "./duration-reopen-tests.js";